refactor(vt-flow): consolidate enums + API option interfaces into types.ts
