--- packages/vt-flow/src/VtFlowRole.ts
@@ -1,5 +1,0 @@
-/** Role a party plays in a vt-flow session; perspective-only, never on the wire. */
-export enum VtFlowRole {
-  Applicant = 'applicant',
-  Validator = 'validator',
-}
